+++ netidx-tools/src/browser/mod.rs
@@ -30,6 +30,8 @@ pub async fn run(
         .await
         .context("creating publisher")?;
     let subscriber = Subscriber::new(cfg, auth).context("create subscriber")?;
+    crate::log_errors::publisher(&publisher);
+    crate::log_errors::subscriber(&subscriber);
     let net_config = NetConfig::Ready { publisher, subscriber };
     ShellBuilder::<NoExt>::default()
         .mode(Mode::Script(Source::Internal(literal!(include_str!("browser.gx")))))