+++ netidx-tools/src/main.rs
@@ -1,5 +1,6 @@
 #![recursion_limit = "2048"]
 mod browser;
+mod log_errors;
 mod publisher;
 mod record_client;
 mod resolver;