+++ netidx-tools/src/subscriber.rs
@@ -478,6 +478,7 @@ mod tests {
 pub(super) async fn run(cfg: Config, auth: DesiredAuth, p: Params) -> Result<()> {
     env_logger::init();
     let subscriber = Subscriber::new(cfg, auth).context("create subscriber")?;
+    crate::log_errors::subscriber(&subscriber);
     let mut ctx = Ctx::new(subscriber, p);
     let mut tick = time::interval(Duration::from_secs(1));
     loop {