+++ netidx-tools/src/wsproxy.rs
@@ -27,6 +27,8 @@ pub(super) async fn run(
                 .await
                 .context("creating publisher")?;
             let subscriber = Subscriber::new(cfg, auth).context("creating subscriber")?;
+            crate::log_errors::publisher(&publisher);
+            crate::log_errors::subscriber(&subscriber);
             Ok((publisher, subscriber))
         }
     };